fix: bad PostLocationDto mapping

+++ AirsoftBmsApp/Model/Dto/Location/PostLocationDto.cs
@@ -5,6 +5,6 @@ public class PostLocationDto
     public double Longitude { get; set; }
     public double Latitude { get; set; }
     public double Accuracy { get; set; }
-    public double Bearing { get; set; }
+    public int Bearing { get; set; }
     public DateTimeOffset Time { get; set; }
 }
